requirements.txt
aiogram==3.13.1
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.2
httpx==0.27.2
pyjwt==2.9.0

